+++ cdap-common/src/test/java/io/cdap/cdap/common/security/HttpsEnablerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018-2019 Cask Data, Inc.
+ * Copyright © 2019 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package io.cdap.cdap.security.tools;
+package io.cdap.cdap.common.security;
 
 import io.cdap.http.AbstractHttpHandler;
 import io.cdap.http.HttpResponder;
