+++ cdap-security/src/test/java/io/cdap/cdap/security/server/ExternalMTLSAuthenticationServerTest.java
@@ -19,8 +19,8 @@ package io.cdap.cdap.security.server;
 import io.cdap.cdap.common.conf.CConfiguration;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.conf.SConfiguration;
+import io.cdap.cdap.common.security.HttpsEnabler;
 import io.cdap.cdap.common.utils.Networks;
-import io.cdap.cdap.security.tools.HttpsEnabler;
 import org.eclipse.jetty.plus.jaas.spi.PropertyFileLoginModule;
 import org.junit.AfterClass;
 import org.junit.Assert;
